(gh-19) Run puppetdb ssl-setup in acceptance pre-suite
The openvoxdb package (and puppetdb) has a post-install script that runs
puppetdb ssl-setup. This will prep openvoxdb's certs by copying the
agent certs, but this only works if the agent has generated certs.

But for the acceptance pipeline, we're installing openvox-agent,
openvox-server and openvoxdb with Bolt before the suite is run, and no
configuration is being done before hand.

The order looks like:

1. kvm_automation_tooling::install_openvox installs the packages.
   The puppetdb ssl-setup run by the package manager does nothing since
   there are no certs yet.
2. Beaker pre-suite/foss steps < 95 setup agent/server certs.
3. Beaker pre-suite/foss step 95 uses the puppetdb module to ensure the
   package is installed (it is per #1) and to configure puppetdb, but
   fails configuring because puppetdb doesn't have certs yet.

By inserting a manual 'puppetdb ssl-setup' in the 95_install_pdb.rb
before the puppetdb classes are enforced we get the certs in place and
the module can proceed.

The ssl-setup command should be idempotent, just in case they are
already in place.