(gh-19) Provide pg repos on el platforms
Test runs could not find postgresql 14 on el-8. This snippet (slightly
modified) comes from the main branch and ensures a pg repo is configured
for any el platform so we have access to postgresql packages.

Seems like this should really be part of the puppetdb module though?
Or we should be using the postgresql module maybe?