Changes for FIPS
This makes several changes to build a FIPS-compatible server. At
Perforce, FIPS is only supported for PE, and with PE, there is a
separate package that installs the Bouncy Castle jars. To avoid having
to have a separate package, this has ezbake package them up in the
server RPM for FIPS platforms.

To do this, the build rake task runs 'lein with-profile fips classpath'.
This causes lein to pull down all of the jars, including the FIPS Bouncy
Castle jars, to spit out the classpath. We then pull those jars directly
from the Maven cache on disk.

We put them in resources/ext/build-scripts/bc-fips-jars. This is a bit
of a hack to have ezbake automatically copy them into the target staging
area, so that we don't need to add another mechanism to ezbake. Then, in
resources/ext/ezbake.conf, we append some install commands so that when
this directory exists, then we do the appropriate 'install' commands to
include them in the rpm.

Additionally, the CLASSPATH for the service is modified to put this
directory with the jars (/opt/puppetlabs/server/apps/puppetserver/jars)
first so the BC FIPS jars are loaded before the rest of the server.

Finally, a java.security.fips file is laid down and the JAVA_ARGS
include this file to enable using the BC FIPS code. Previously, we had
separate versions of these files for different Java versions. This isn't
really necessary with Java 17 and 21, so we just provide a single file.